backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.2
pydantic==2.5.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
resend==0.7.2
emails==0.6
jinja2==3.1.2
